Integrate api

+++ lib/core/providers/auth_provider.dart
@@ -1,14 +1,37 @@
 import 'package:flutter/cupertino.dart';
-import 'package:pedo/core/services/auth_service.dart';
+import 'package:pedo/config/endpoint.dart';
+import 'package:pedo/core/models/user_model.dart';
 import 'package:pedo/utils/secure_storage_service.dart';
 
 class AuthProvider with ChangeNotifier {
+  bool isLoading = false;
+  UserModel? setUser;
+  UserModel? get getUser => setUser;
+
   // Login
   Future login(String email, String password) async {
-    try {
-      var response = await AuthService.login(email, password);
+    isLoading = true;
+    notifyListeners();
 
-      return response;
+    try {
+      var response = await Endpoint.urlLogin(
+        body: {
+          'email': email,
+          'password': password,
+        },
+      );
+
+      if (response['code'] == 200) {
+        await SecureStorageService.setToken(response['data']['token']);
+        setUser = UserModel.fromJson(response['data']['user'])
+          ..token = response['data']['token'];
+        isLoading = false;
+
+        notifyListeners();
+        return response;
+      } else {
+        return response;
+      }
     } catch (err) {
       debugPrint(err.toString());
     }
@@ -16,20 +39,34 @@ class AuthProvider with ChangeNotifier {
 
   // Login With Token
   Future loginWithToken() async {
-    String token = await SecureStorageService.getToken();
-
     try {
-      var response = await AuthService.loginWithToken(token);
+      String token = await SecureStorageService.getToken();
+      var response = await Endpoint.urlProfile();
 
-      return response;
+      if (response['code'] == 200) {
+        setUser = UserModel.fromJson(response['data']['user'])..token = token;
+        isLoading = false;
+
+        notifyListeners();
+        return response;
+      } else {
+        SecureStorageService.removeToken();
+        return response;
+      }
     } catch (err) {
       debugPrint(err.toString());
     }
   }
 
+// Register
   Future register(String name, String email, String password) async {
     try {
-      var response = await AuthService.register(name, email, password);
+      Map<String, dynamic> body = {
+        'name': name,
+        'email': email,
+        'password': password,
+      };
+      var response = await Endpoint.urlRegister(body);
 
       return response;
     } catch (err) {
@@ -40,10 +77,17 @@ class AuthProvider with ChangeNotifier {
   // Logout
   Future logout() async {
     try {
-      var token = await SecureStorageService.getToken();
-      SecureStorageService.removeToken();
+      var response = await Endpoint.urlLogout();
+      if (response['code'] == 200) {
+        SecureStorageService.removeToken();
+        setUser = null;
+        isLoading = false;
 
-      await AuthService.logout(token.toString());
+        notifyListeners();
+        return response;
+      } else {
+        return response;
+      }
     } catch (err) {
       debugPrint(err.toString());
     }
